+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Support for Archive Log Backup and Point in time restore
 - Support resource discovery for `datacatalog` resources
 
+### Fixed
+- Fixed plan failure in case of missing required attributes in resource discovery. Placeholder values will be added for missing required attributes and the attributes will be added to `lifecycle ignore_changes`
+ 
 ## 3.80.0 (June 10, 2020)
 
 ### Added